+++ azure-cli.pyproj
@@ -118,6 +118,9 @@
     <Compile Include="command_modules\azure-cli-acs\azure\cli\command_modules\acs\__init__.py" />
     <Compile Include="command_modules\azure-cli-acs\setup.py" />
     <Compile Include="command_modules\azure-cli-acs\tests\test_acs_client.py" />
+    <Compile Include="command_modules\azure-cli-acs\tests\test_acs_commands.py">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="command_modules\azure-cli-acs\tests\test_custom.py" />
     <Compile Include="command_modules\azure-cli-acs\tests\test_service_principals.py" />
     <Compile Include="command_modules\azure-cli-acs\tests\__init__.py" />
