+++ src/command_modules/azure-cli-acs/tests/recordings/test_acs_create_update.yaml
